+++ pytesmo/io/dataset_base.py
